Struct solana_banks_server::banks_server::BanksServer

struct BanksServer {
    bank_forks: Arc<RwLock<BankForks>>,
    block_commitment_cache: Arc<RwLock<BlockCommitmentCache>>,
    transaction_sender: Sender<TransactionInfo>,
    poll_signature_status_sleep_duration: Duration,
}

Fields
bank_forks: Arc<RwLock<BankForks>>block_commitment_cache: Arc<RwLock<BlockCommitmentCache>>transaction_sender: Sender<TransactionInfo>poll_signature_status_sleep_duration: Duration
Implementations

impl BanksServer

fn new( bank_forks: Arc<RwLock<BankForks>>, block_commitment_cache: Arc<RwLock<BlockCommitmentCache>>, transaction_sender: Sender<TransactionInfo>, poll_signature_status_sleep_duration: Duration ) -> Self

Return a BanksServer that forwards transactions to the given sender. If unit-testing, those transactions can go to a bank in the given BankForks. Otherwise, the receiver should forward them to a validator in the leader schedule.

fn run( bank_forks: Arc<RwLock<BankForks>>, transaction_receiver: Receiver<TransactionInfo> )

fn new_loopback( bank_forks: Arc<RwLock<BankForks>>, block_commitment_cache: Arc<RwLock<BlockCommitmentCache>>, poll_signature_status_sleep_duration: Duration ) -> Self

Useful for unit-testing

fn slot(&self, commitment: CommitmentLevel) -> Slot

fn bank(&self, commitment: CommitmentLevel) -> Arc<Bank>

async fn poll_signature_status( self, signature: &Signature, blockhash: &Hash, last_valid_block_height: u64, commitment: CommitmentLevel ) -> Option<Result<()>>